RAS: WORD loop counters against DWORD header fields never terminated
Load() declared

    WORD linelength;
    WORD fill;
    WORD x, y;

but header.width and header.height are DWORDs straight out of the file, and
they are what FreeImage_AllocateHeader() is given. In

    for (y = 0; y < header.height; y++) {
            ReadData(io, handle, bits, linelength, rle);
            bits -= pitch;
            ...
    }

the unsigned short promotes to int and then converts to unsigned, so for a
height above 65535 the condition is always true while y wraps at 65536: the
row loop never ends, and bits keeps walking down past the front of the
allocation. With RT_BYTE_ENCODED the RLE branch of ReadData() writes whatever
it last held for every byte it is asked for, so that is an unbounded heap
underflow; with RT_STANDARD nothing is written and it is simply an endless
loop. The column loops had the same defect for widths above 65535, running
bits and bp off the ends of both the scanline and the source buffer.

Widening the four to unsigned matches the fields they are compared against.
That also stops linelength truncating: at WORD it wrapped for any 8-bit image
wider than 65535, which read every row short.

Confirmed: 1 x 70000 RT_BYTE_ENCODED gave an ASan heap-buffer-overflow in
ReadData and RT_STANDARD hung; 100000 x 1 at 24 bpp gave a UBSan report for a
read past the line buffer. All four now decode at their declared size with no
sanitizer output.

Note the dimensions still reach FreeImage_AllocateHeader() as int, so a width
or height above INT_MAX is a separate problem, handled with the abs() change
in BitmapAccess.

Finding 5 of AUDIT.md.

+++ Source/FreeImage/PluginRAS.cpp
@@ -193,15 +193,18 @@ SupportsNoPixels() {
 static FIBITMAP * DLL_CALLCONV
 Load(FreeImageIO *io, fi_handle handle, int page, int flags, void *data) {
 	SUNHEADER header;	// Sun file header
-	WORD linelength;	// Length of raster line in bytes
-	WORD fill;			// Number of fill bytes per raster line
+	// these four track DWORD header fields: as WORDs they truncated the line
+	// length and, worse, made the row and column loops below wrap at 65536 and
+	// never terminate
+	unsigned linelength;	// Length of raster line in bytes
+	unsigned fill;			// Number of fill bytes per raster line
 	BOOL rle;			// TRUE if RLE file
 	BOOL isRGB;			// TRUE if file type is RT_FORMAT_RGB
 	BYTE fillchar;
 
 	FIBITMAP *dib = NULL;
 	BYTE *bits;			// Pointer to dib data
-	WORD x, y;
+	unsigned x, y;
 
 	if(!handle) {
 		return NULL;
@@ -364,9 +367,9 @@ Load(FreeImageIO *io, fi_handle handle, int page, int flags, void *data) {
 		// Each row is multiple of 16 bits (2 bytes).
 
 		if (header.depth == 1) {
-			linelength = (WORD)((header.width / 8) + (header.width % 8 ? 1 : 0));
+			linelength = (header.width / 8) + (header.width % 8 ? 1 : 0);
 		} else {
-			linelength = (WORD)header.width;
+			linelength = header.width;
 		}
 
 		fill = (linelength % 2) ? 1 : 0;
